feat(6895): extract group channel crypto modules
