+++ crates/kamn-core/src/group_channel_crypto/crypto_helpers.rs
@@ -1,0 +1,16 @@
+mod derivation;
+mod encoding;
+mod nonce;
+mod signature;
+mod zeroize_support;
+
+pub(crate) use derivation::{
+    derive_group_aead_key, derive_group_aead_key_legacy, derive_group_shared_secret,
+};
+pub(crate) use encoding::{hex_decode, hex_encode};
+pub(crate) use nonce::{group_nonce_bytes, legacy_raw_prefix_group_nonce_bytes};
+pub(crate) use signature::compute_signature;
+pub(crate) use zeroize_support::{
+    zeroize_sender_key_history,
+    zeroize_u64_keyed_sender_history,
+};
