+++ crates/kamn-core/src/group_channel_crypto/tests.rs
@@ -1,0 +1,4 @@
+mod decrypt_contract_tests;
+mod lifecycle_contract_tests;
+mod regression_contract_tests;
+mod support;
